manyhow

Struct ErrorMessage

pub struct ErrorMessage { /* private fields */ }

A single error message

Can take additional attachments like help or note.

Implements ToTokensError and can therefore be used with MacroOutputs.

Implementations

impl ErrorMessage

pub fn new(span: impl SpanRanged, msg: impl Display) -> Self

Creates a new error message at the specified span

This function takes a SpanRanged meaning you can also pass a Range<Span> (i.e. first..last) for better error messages on multi token values, for details see SpanRanged#motivation

If your type implements ToTokens use ErrorMessage::spanned instead.

pub fn spanned(tokens: impl ToTokens, msg: impl Display) -> Self

Creates an error message pointing to the complete token stream tokens expands to

pub fn call_site(msg: impl Display) -> Self

Creates a new error message at Span::call_site prefer ErrorMessage::new or ErrorMessage::spanned with the correct span for a more helpful output.

pub fn attachment(self, label: &'static str, msg: impl Display) -> Self

Attaches an additional message to self reusing the same span, and the specified label.

pub fn error(self, msg: impl Display) -> Self

Attaches a new error message to self reusing the same span

pub fn warning(self, msg: impl Display) -> Self

Attaches a new warning message to self reusing the same span

pub fn note(self, msg: impl Display) -> Self

Attaches a new note message to self reusing the same span

pub fn help(self, msg: impl Display) -> Self

Attaches a new help message to self reusing the same span