Trait manyhow::MacroHandler

pub trait MacroHandler<Input, Output, Function, Error = Infallible> {
    // Required method
    fn call(
        self,
        input: Input,
        dummy: &mut Output,
        emitter: &mut Emitter
    ) -> Result<Output, Error>;
}

Implementation of a proc-macro

Note: for TokenStream either proc_macro::TokenStream or proc_macro2::TokenStream can be used.

Trait is implemented for any function, taking in either one (for derive or function like macros) or two (for attribute macros) TokenStreams and returns either a TokenStream or a Result<TokenStream, impl ToTokensError>.

Additionally they can take optionally in any order a &mut Emitter which allows emitting errors without returning early. And a &mut TokenStream to return a dummy TokenStream on failure, note that this TokenStream must be the same type as the one returned.

Required Methods

fn call( self, input: Input, dummy: &mut Output, emitter: &mut Emitter ) -> Result<Output, Error>